Boostbook: Regenerate the "gold" test docs.
The ids are now a little more truncated than they used to be, this is partly
due to the fix in r77599, and partly due to the generate-id generating longer
ids in newer versions of xsltproc.

It might be better to improve the id normalisation so that it is more
consistent across versions.

git-svn-id: https://svn.boost.org/svn/boost/trunk@82891 b8fc166d-592f-0410-95f2-cb63ce0dd405